+++ src/main/java/roomescape/dao/ReservationTimeDao.java
@@ -1,15 +1,58 @@
 package roomescape.dao;
 
+import java.time.LocalTime;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import javax.sql.DataSource;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
 import roomescape.domain.ReservationTime;
+import roomescape.domain.ReservationTimeRepository;
 
-public interface ReservationTimeDao {
-    List<ReservationTime> findAll();
+@Repository
+public class ReservationTimeDao implements ReservationTimeRepository {
+    private final JdbcTemplate jdbcTemplate;
+    private final SimpleJdbcInsert jdbcInsert;
+    private final RowMapper<ReservationTime> timeRowMapper = (resultSet, rowNum) -> new ReservationTime(
+            resultSet.getLong("id"),
+            LocalTime.parse(resultSet.getString("start_at"))
+    );
 
-    Optional<ReservationTime> findById(Long id);
+    public ReservationTimeDao(JdbcTemplate jdbcTemplate, DataSource dataSource) {
+        this.jdbcTemplate = jdbcTemplate;
+        this.jdbcInsert = new SimpleJdbcInsert(dataSource)
+                .withTableName("reservation_time")
+                .usingGeneratedKeyColumns("id");
+    }
 
-    void save(ReservationTime time);
+    @Override
+    public List<ReservationTime> findAll() {
+        String sql = "SELECT * FROM reservation_time";
+        return jdbcTemplate.query(sql, timeRowMapper);
+    }
 
-    void delete(ReservationTime time);
+    @Override
+    public Optional<ReservationTime> findById(Long id) {
+        String sql = "SELECT * FROM reservation_time WHERE id = ?";
+        ReservationTime time = jdbcTemplate.queryForObject(sql, timeRowMapper, id);
+        return Optional.ofNullable(time);
+    }
+
+    @Override
+    public void save(ReservationTime time) {
+        Map<String, Object> parameters = new HashMap<>();
+        parameters.put("start_at", time.getStartAt());
+        Long id = (Long) jdbcInsert.executeAndReturnKey(parameters);
+        time.setId(id);
+    }
+
+    @Override
+    public void delete(ReservationTime time) {
+        String sql = "DELETE FROM reservation_time WHERE id = ?";
+        jdbcTemplate.update(sql, time.getId());
+    }
 }
