--- src/main/java/roomescape/dao/ReservationTimeDaoImpl.java
@@ -1,57 +1,0 @@
-package roomescape.dao;
-
-import java.time.LocalTime;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import javax.sql.DataSource;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
-import roomescape.domain.ReservationTime;
-
-@Repository
-public class ReservationTimeDaoImpl implements ReservationTimeDao {
-    private final JdbcTemplate jdbcTemplate;
-    private final SimpleJdbcInsert jdbcInsert;
-    private final RowMapper<ReservationTime> timeRowMapper = (resultSet, rowNum) -> new ReservationTime(
-            resultSet.getLong("id"),
-            LocalTime.parse(resultSet.getString("start_at"))
-    );
-
-    public ReservationTimeDaoImpl(JdbcTemplate jdbcTemplate, DataSource dataSource) {
-        this.jdbcTemplate = jdbcTemplate;
-        this.jdbcInsert = new SimpleJdbcInsert(dataSource)
-                .withTableName("reservation_time")
-                .usingGeneratedKeyColumns("id");
-    }
-
-    @Override
-    public List<ReservationTime> findAll() {
-        String sql = "SELECT * FROM reservation_time";
-        return jdbcTemplate.query(sql, timeRowMapper);
-    }
-
-    @Override
-    public Optional<ReservationTime> findById(Long id) {
-        String sql = "SELECT * FROM reservation_time WHERE id = ?";
-        ReservationTime time = jdbcTemplate.queryForObject(sql, timeRowMapper, id);
-        return Optional.ofNullable(time);
-    }
-
-    @Override
-    public void save(ReservationTime time) {
-        Map<String, Object> parameters = new HashMap<>();
-        parameters.put("start_at", time.getStartAt());
-        Long id = (Long) jdbcInsert.executeAndReturnKey(parameters);
-        time.setId(id);
-    }
-
-    @Override
-    public void delete(ReservationTime time) {
-        String sql = "DELETE FROM reservation_time WHERE id = ?";
-        jdbcTemplate.update(sql, time.getId());
-    }
-}
