+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -2,21 +2,21 @@ package roomescape.service;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.dao.ReservationTimeDao;
 import roomescape.domain.ReservationTime;
+import roomescape.domain.ReservationTimeRepository;
 import roomescape.web.dto.ReservationTimeRequest;
 import roomescape.web.dto.ReservationTimeResponse;
 
 @Service
 public class ReservationTimeService {
-    private final ReservationTimeDao reservationTimeDao;
+    private final ReservationTimeRepository reservationTimeRepository;
 
-    public ReservationTimeService(ReservationTimeDao reservationTimeDao) {
-        this.reservationTimeDao = reservationTimeDao;
+    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+        this.reservationTimeRepository = reservationTimeRepository;
     }
 
     public List<ReservationTimeResponse> findAllReservationTime() {
-        List<ReservationTime> reservationTimes = reservationTimeDao.findAll();
+        List<ReservationTime> reservationTimes = reservationTimeRepository.findAll();
         return reservationTimes.stream()
                 .map(ReservationTimeResponse::from)
                 .toList();
@@ -24,17 +24,17 @@ public class ReservationTimeService {
 
     public ReservationTimeResponse saveReservationTime(ReservationTimeRequest request) {
         ReservationTime reservationTime = request.toReservationTime();
-        reservationTimeDao.save(reservationTime);
+        reservationTimeRepository.save(reservationTime);
         return ReservationTimeResponse.from(reservationTime);
     }
 
     public void deleteReservationTime(Long id) {
         ReservationTime reservationTime = findReservationTimeById(id);
-        reservationTimeDao.delete(reservationTime);
+        reservationTimeRepository.delete(reservationTime);
     }
 
     private ReservationTime findReservationTimeById(Long id) {
-        return reservationTimeDao.findById(id)
+        return reservationTimeRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 시간입니다."));
     }
 }
